age-gender-image-api/requirements.txt
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
cycler==0.10.0
dlib==19.9.0
enum34==1.1.6
Flask==0.12.2
html5lib==0.9999999
idna==2.6
imutils==0.4.5
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.0
numpy==1.14.2
opencv-python==3.4.0.12
pandas==0.22.0
Pillow==5.0.0
pkg-resources==0.0.0
protobuf==3.5.2
pyparsing==2.2.0
python-dateutil==2.7.0
pytz==2018.3
requests==2.18.4
six==1.11.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
gunicorn==19.8.1
